+++ proxy/handler.go
@@ -16,7 +16,7 @@ import (
 
 var re = regexp.MustCompile(`^(http:|https:)?/?/?(.*)`) // 匹配http://或https://开头的路径
 
-func NoRouteHandler(cfg *config.Config, limiter *rate.RateLimiter, iplimiter *rate.IPRateLimiter, runMode string) app.HandlerFunc {
+func NoRouteHandler(cfg *config.Config, limiter *rate.RateLimiter, iplimiter *rate.IPRateLimiter) app.HandlerFunc {
 	return func(ctx context.Context, c *app.RequestContext) {
 
 		// 限制访问频率
@@ -35,14 +35,12 @@ func NoRouteHandler(cfg *config.Config, limiter *rate.RateLimiter, iplimiter *ra
 			}
 
 			if !allowed {
-				//c.JSON(http.StatusTooManyRequests, gin.H{"error": "Too Many Requests"})
 				c.JSON(http.StatusTooManyRequests, map[string]string{"error": "Too Many Requests"})
 				logWarning("%s %s %s %s %s 429-TooManyRequests", c.ClientIP(), c.Request.Method, c.Request.RequestURI(), c.Request.Header.UserAgent(), c.Request.Header.GetProtocol())
 				return
 			}
 		}
 
-		//rawPath := strings.TrimPrefix(c.Request.URL.Path, "/") // 去掉前缀/
 		rawPath := strings.TrimPrefix(string(c.Request.RequestURI()), "/") // 去掉前缀/
 		matches := re.FindStringSubmatch(rawPath)                          // 匹配路径
 		logInfo("Matches: %v", matches)
@@ -75,18 +73,16 @@ func NoRouteHandler(cfg *config.Config, limiter *rate.RateLimiter, iplimiter *ra
 
 		logInfo("%s %s %s %s %s Matched-Username: %s, Matched-Repo: %s", c.ClientIP(), c.Request.Method, rawPath, c.Request.Header.UserAgent(), c.Request.Header.GetProtocol(), username, repo)
 		// dump log 记录详细信息 c.ClientIP(), c.Request.Method, rawPath,c.Request.Header.UserAgent(), c.Request.Header.GetProtocol(), full Header
-		logDump("%s %s %s %s %s %s", c.ClientIP(), c.Request.Method, rawPath, c.Request.Header.UserAgent(), c.Request.Header.GetProtocol(), c.Request.Header)
+		logDump("%s %s %s %s %s %s", c.ClientIP(), c.Request.Method, rawPath, c.Request.Header.UserAgent(), c.Request.Header.GetProtocol(), c.Request.Header.Header())
 		repouser := fmt.Sprintf("%s/%s", username, repo)
 
 		// 白名单检查
 		if cfg.Whitelist.Enabled {
 			whitelist := auth.CheckWhitelist(username, repo)
 			if !whitelist {
-				logErrMsg := fmt.Sprintf("%s %s %s %s %s Whitelist Blocked repo: %s", c.ClientIP(), c.Request.Method, rawPath, c.Request.Header.UserAgent(), c.Request.Header.GetProtocol(), repouser)
 				errMsg := fmt.Sprintf("Whitelist Blocked repo: %s", repouser)
-				//c.JSON(http.StatusForbidden, gin.H{"error": errMsg})
 				c.JSON(http.StatusForbidden, map[string]string{"error": errMsg})
-				logWarning(logErrMsg)
+				logWarning("%s %s %s %s %s Whitelist Blocked repo: %s", c.ClientIP(), c.Request.Method, rawPath, c.Request.Header.UserAgent(), c.Request.Header.GetProtocol(), repouser)
 				return
 			}
 		}
@@ -95,23 +91,12 @@ func NoRouteHandler(cfg *config.Config, limiter *rate.RateLimiter, iplimiter *ra
 		if cfg.Blacklist.Enabled {
 			blacklist := auth.CheckBlacklist(username, repo)
 			if blacklist {
-				logErrMsg := fmt.Sprintf("%s %s %s %s %s Blacklist Blocked repo: %s", c.ClientIP(), c.Request.Method, rawPath, c.Request.Header.UserAgent(), c.Request.Header.GetProtocol(), repouser)
 				errMsg := fmt.Sprintf("Blacklist Blocked repo: %s", repouser)
-				//c.JSON(http.StatusForbidden, gin.H{"error": errMsg})
 				c.JSON(http.StatusForbidden, map[string]string{"error": errMsg})
-				logWarning(logErrMsg)
+				logWarning("%s %s %s %s %s Blacklist Blocked repo: %s", c.ClientIP(), c.Request.Method, rawPath, c.Request.Header.UserAgent(), c.Request.Header.GetProtocol(), repouser)
 				return
 			}
 		}
-
-		/*
-			matches = CheckURL(rawPath, c)
-			if matches == nil {
-				c.AbortWithStatus(http.StatusNotFound)
-				logWarning("%s %s %s %s %s 404-NOMATCH", c.ClientIP(), c.Request.Method, rawPath,c.Request.Header.UserAgent(), c.Request.Header.GetProtocol())
-				return
-			}
-		*/
 
 		// 若匹配api.github.com/repos/用户名/仓库名/路径, 则检查是否开启HeaderAuth
 
@@ -137,8 +122,7 @@ func NoRouteHandler(cfg *config.Config, limiter *rate.RateLimiter, iplimiter *ra
 		case "releases", "blob", "raw", "gist", "api":
 			ChunkedProxyRequest(ctx, c, rawPath, cfg, matcher)
 		case "clone":
-			//ProxyRequest(c, rawPath, cfg, "git", runMode)
-			GitReq(ctx, c, rawPath, cfg, "git", runMode)
+			GitReq(ctx, c, rawPath, cfg, "git")
 		default:
 			c.String(http.StatusForbidden, "Invalid input.")
 			fmt.Println("Invalid input.")
@@ -146,16 +130,3 @@ func NoRouteHandler(cfg *config.Config, limiter *rate.RateLimiter, iplimiter *ra
 		}
 	}
 }
-
-/*
-func CheckURL(u string, c *gin.Context) []string {
-	for _, exp := range exps {
-		if matches := exp.FindStringSubmatch(u); matches != nil {
-			return matches[1:]
-		}
-	}
-	errMsg := fmt.Sprintf("%s %s %s %s %s Invalid URL", c.ClientIP(), c.Request.Method, u,c.Request.Header.UserAgent(), c.Request.Header.GetProtocol())
-	logError(errMsg)
-	return nil
-}
-*/
